:recycle:  refactor: joinUsers 타입 Integer -> Long 변경 [#34]

+++ waggle-api/src/main/java/com/trip/api/chat/dto/request/CreateChatRoomRequest.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 public class CreateChatRoomRequest {
 
     @NotEmpty
-    private ArrayList<Integer> joinUsers;
+    private ArrayList<Long> joinUsers;
 
     @NotNull
     private Long chatRoomMaker;
